requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
pinecone-client>=3.0.0
pypdf2>=3.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langchain-pinecone>=0.1.0
langchain-community>=0.0.10
openai>=1.0.0
tiktoken>=0.5.0
